Split out socket BIO tests
